requirements.txt
openpyxl>=3.1.0
python-docx>=1.1.0
pytest>=8.0.0
pypdf>=4.0.0
